+++ src/library/states/states.c
@@ -900,30 +900,6 @@ void states_new_iteration(int my_id, uint period, uint iterations,
             /* We must evaluate policy decissions */
             signatures_stables++;
 
-            if (default_signatures_different(&loop_signature, &policy_sel_signature, 0.2)) {
-
-              /* Reset */
-              policy_set_default_freq(&loop_signature);
-              EAR_STATE = EVALUATING_LOCAL_SIGNATURE;
-              /* And then we mark as not ready */
-              if (master) clean_signatures(lib_shared_region,sig_shared_region);
-
-              if (VERB_ON(2)) {
-
-                char sigs_metrics_buff[128]; // This buffer is used to show differences in signatures.
-
-                snprintf(sigs_metrics_buff, sizeof(sigs_metrics_buff),
-                    "(CPI/GB/s/DC) current (%.2lf/%.2lf/%.2lf) policy (%.2lf/%.2lf/%.2lf)",
-                    loop_signature.CPI, loop_signature.GBS, loop_signature.DC_power,
-                    policy_sel_signature.CPI, policy_sel_signature.GBS, policy_sel_signature.DC_power);
-
-                verbose_info_master("EARL state reset because signature changed from the last passed to the policy: %s", sigs_metrics_buff);
-              }
-
-              overhead_stop(id_ovh_stable);
-              break;
-            }
-
             if (sig_ready[def_pstate] == 0) {
                 verbose_info2_master("Signature at default freq not available, assuming policy ok.");
                 /* If default is not available, that means a dynamic configuration has been decided, we assume we are ok */