+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### EAR 5.1.4
+- Change the open file limit in edcmon.
+- Increase the timeout for remote calls.
+- Don't check for socket alive to reduce remote message parsing overhead.
+- Add new signature classification changes threshold for AMD zen3.
+- eUFS and MPI Load Balance disabled by default.
+- Bugs fixed:
+  - Minor fixes in signature different decision algorithm.
+  - Dimension in hsmp structure fixed.
+
 ### EAR 5.1.3
 - Bugs fixed:
   - Prevent closing remote connection.